+++ eapli.expensemanager/src/Persistance/JPA/PaymentMeansJPA.java
@@ -20,7 +20,7 @@ import javax.persistence.Query;
  *
  * @author Catia Marques 1120066
  */
-public class PaymentMeansJPA extends JpaGeneric<PaymentMean, Long> implements IPaymentMeansRepository {
+public class PaymentMeansJPA extends JpaGeneric<PaymentMean, Integer> implements IPaymentMeansRepository {
 
     @Override
     public void registerPaymentMean(PaymentMean pm) {
